// src/app/models/editUser.ts
export interface EditedUser {
    id: number;
    name: string;
    email: string;
    role: string;
    leaveType: number;
    address:string;
    joinDate: string;
    employeeId: string;
    designation: string;
    cnic: string;
    contactNo: string;
    famContactNo: string;
    image: any;
    schedule:number;
    
  }